Rename timestamp_cleanup to standardize_timestamp
For the 'Case Study 1: How does a bike-share navigate speedy success?' Capstone project for the Google Data Analytics Professional Certificate project. Partnership with Divvy bike-sharing company using historical service data.

The trip date and time data varies in format from year to year, so I wanted to standardize it in a way that would be easy for BigQuery SQL environment to read it.