Create standardize_station-names
For the 'Case Study 1: How does a bike-share navigate speedy success?' Capstone project for the Google Data Analytics Professional Certificate project. Partnership with Divvy bike-sharing company using historical service data.

This is a problem file because it does not solve the issue. When standardizing the Station columns, I found there was a lot of variation. Some of these transformations actually belong in the station_id column, not the Station names columns.