Create standardize_rider-bday
For the 'Case Study 1: How does a bike-share navigate speedy success?' Capstone project for the Google Data Analytics Professional Certificate project. Partnership with Divvy bike-sharing company using historical service data.

the user birthday column, which is not used in most years, became troublesome. This was an attempt to clean that up.